+++ src/main/java/hanium/dtc/auth/controller/KakaoLoginPageController.java
@@ -1,4 +1,4 @@
-package hanium.dtc.security.controller;
+package hanium.dtc.auth.controller;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
